refactor(server): consolidate pagination parsing into shared helper
Moves paginationFromQuery from playlist_handlers.go into a dedicated
pagination.go and updates 7 duplicate hand-rolled parsers to use it.
Eliminates inconsistent limit/offset handling. Structure audit STRUCT-2.
