+++ internal/server/metadata_batch_candidates.go
@@ -1,5 +1,5 @@
 // file: internal/server/metadata_batch_candidates.go
-// version: 1.8.1
+// version: 1.8.2
 // guid: a1b2c3d4-e5f6-7a8b-9c0d-e1f2a3b4c5d6
 // last-edited: 2026-05-01
 
@@ -11,7 +11,6 @@ import (
 	"fmt"
 	"log"
 	"net/http"
-	"strconv"
 	"sync"
 	"sync/atomic"
 	"time"
@@ -328,18 +327,7 @@ func (s *Server) handleGetOperationResults(c *gin.Context) {
 		return
 	}
 
-	limit := 250
-	if raw := c.Query("limit"); raw != "" {
-		if n, err := strconv.Atoi(raw); err == nil && n >= 0 {
-			limit = n
-		}
-	}
-	offset := 0
-	if raw := c.Query("offset"); raw != "" {
-		if n, err := strconv.Atoi(raw); err == nil && n >= 0 {
-			offset = n
-		}
-	}
+	limit, offset := paginationFromQuery(c)
 
 	store := s.Store()
